test ml v3

+++ RaspberryPi/Consumer/Consumer.py
@@ -64,7 +64,7 @@ def post_to_predict(data):
 # Function to write data to InfluxDB
 def write_to_influxdb(data):
     # Convert g8_sensor_data to an integer if necessary
-    temp_BMP280 =  int(data["temp_BMP280"])
+    temp_BMP280 =  float(data["temp_BMP280"])
     # temp_HTS221 = int(data["temp_HTS221"])
     # humid_hts221 = int(data["humid_HTS221"])
     # pressure_bmp280 = int(data["pressure_BMP280"])
